requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
# psycopg2-binary==2.9.7  # Comentado - não necessário para SQLite
Pillow==10.0.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-filter==23.3
google-auth>=2.0.0
